aroti-backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
pyjwt==2.8.0
cryptography==41.0.7
httpx==0.25.2

# Redis
redis[hiredis]==5.0.1

# Temporal
temporalio==1.5.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
